+++ gui/src/elements/database/dialog/__init__.py
@@ -2,8 +2,7 @@ from .advanced_search_dialog import _AdvancedSearchDialog
 from .associated_content_dialog import _AssociatedContentDialog
 from .associated_entities_dialog import _AssociatedEntitiesDialog
 from .credit_dialog import _CreditDialog
-from .directory_import_dialog import _DirectoryImportDialog
-from .entity_directory_import_dialog import _EntityDirectoryImportDialog
+from .directory_import import _DirectoryImportDialog, _EntityDirectoryImportDialog
 from .episode_dialog import _EpisodeDialog
 from .recommendation_dialog import _RecommendationDialog